Faet: 프로토콜01 Temp 소수점 처리 (#1)

+++ polihealth-sdk-android-v2/src/main/java/kr/co/hconnect/polihealth_sdk_android_v2/api/daily/model/LTMModel.kt
@@ -17,7 +17,7 @@ data class LTMModel(
     @Serializable
     data class SkinTemp(
         val time: String,
-        val skinTemp: Int
+        val skinTemp: Float
     )
 
     @Serializable
